pyproject.toml
[project]
name = "mcp-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastmcp>=2.11.2",
    "google-generativeai>=0.3.0",
    "markdownify>=1.1.0",
    "pillow>=11.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0"
]